Migrate url.sanitized to BitwardenKit
