+++ BitwardenKit/Core/Platform/Extensions/Error+Networking.swift
@@ -1,4 +1,3 @@
-import BitwardenKit
 import Foundation
 
 public extension Error {
